HDDS-9943. TokenRenewer should close OzoneClient after use (#5809)
