+++ hadoop-ozone/ozonefs/src/main/resources/META-INF/services/org.apache.hadoop.security.token.TokenRenewer
@@ -16,5 +16,4 @@
 # limitations under the License.
 #
 
-org.apache.hadoop.fs.ozone.BasicOzoneClientAdapterImpl$Renewer
-org.apache.hadoop.fs.ozone.BasicRootedOzoneClientAdapterImpl$Renewer
+org.apache.hadoop.fs.ozone.OzoneDelegationTokenRenewer
